pyproject.toml
[project]
name = "masato"
version = "1.1.0"
description = "Modular Amplicon Sequencing Analysis TOokit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biom-format>=2.1.16",
    "biopython>=1.84",
    "ipykernel>=6.29.5",
    "ipython>=8.27.0",
    "ipywidgets>=8.1.5",
    "loky>=3.4.1",
    "pandas>=2.2.3",
    "pytest>=8.3.3",
    "pytest-console-scripts>=1.4.1",
    "rich>=13.8.1",
    "scikit-bio>=0.6.2",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "tqdm>=4.66.5",
    "umi-tools>=1.1.5",
    "dendropy>=5.0.1",
    "matplotlib>=3.9.2",
    "pyarrow>=17.0.0",
    "pyyaml>=6.0.2",
    "polars-lts-cpu>=1.8.2",
]

[project.scripts]
"analysis.py" = "masato.analysis:main"
"analysis_sintax.py" = "masato.analysis_sintax:main"
"get_abundance.py" = "masato.get_abundance:main"
"get_tree.py" = "masato.get_tree:main"
"isolate_utils.py" = "masato.isolate_utils:main"
"plot.py" = "masato.plot:main"
"plot_dm.py" = "masato.plot_dm:main"
"process_nbc_tax.py" = "masato.process_nbc_tax:main"
"process_rrndb.py" = "masato.process_rrndb:main"
"trim.py" = "masato.trim:main"
"usearch_workflow.py" = "masato.usearch_workflow:main"
"utils_blast.py" = "masato.utils_blast:main"
"recording.py" = "masato.read_processer.recording:main"
"run_rdp_classifier.py" = "masato.run_rdp_classifier:main"
# setup scripts
masato_conda_install = "masato.setup.conda_install:main"
masato_download_rdp = "masato.setup.download_rdp:main"
masato_download_test_data = "masato.setup.download_test_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
